Clean delegate factories
