+++ src/GraphQL/ValueResolution/DelegateResolverFactory.cs
@@ -1,50 +1,10 @@
-﻿using System.Collections.Concurrent;
-using System.Diagnostics;
-using System.Linq.Expressions;
+﻿using System.Linq.Expressions;
 using System.Reflection;
-using Microsoft.Extensions.DependencyInjection;
 
 namespace Tanka.GraphQL.ValueResolution;
 
-public static class DelegateResolverFactory
+public class DelegateResolverFactory : DelegateFactoryBase<ResolverContext, Resolver>
 {
-    private static readonly ParameterExpression ContextParam = Expression.Parameter(typeof(ResolverContext), "context");
-
-    private static readonly IReadOnlyDictionary<string, Expression> ContextParamProperties = typeof(ResolverContext)
-        .GetProperties(BindingFlags.Public | BindingFlags.Instance)
-        .ToDictionary(p => p.Name.ToLowerInvariant(), p => (Expression)Expression.Property(ContextParam, p));
-
-
-    private static readonly ConcurrentDictionary<Delegate, Resolver> Cache = new();
-
-
-    private static readonly MethodInfo ThrowMethod = typeof(DelegateResolverFactory)
-        .GetMethod(nameof(Throw), BindingFlags.NonPublic | BindingFlags.Static)!;
-
-    private static readonly MethodInfo GetArgumentMethod = typeof(ArgumentsResolverContextExtensions)
-        .GetMethod(
-            nameof(ArgumentsResolverContextExtensions.GetArgument),
-            new[] { typeof(ResolverContextBase), typeof(string) }
-        )!;
-
-    private static readonly MethodInfo BindInputObjectMethod = typeof(ArgumentsResolverContextExtensions)
-        .GetMethod(
-            nameof(ArgumentsResolverContextExtensions.BindInputObject),
-            new[] { typeof(ResolverContextBase), typeof(string) }
-        )!;
-
-    private static readonly MethodInfo BindInputObjectListMethod = typeof(ArgumentsResolverContextExtensions)
-        .GetMethod(
-            nameof(ArgumentsResolverContextExtensions.BindInputObjectList),
-            new[] { typeof(ResolverContextBase), typeof(string) }
-        )!;
-
-    private static readonly MethodInfo HasArgumentMethod = typeof(ArgumentsResolverContextExtensions)
-        .GetMethod(
-            nameof(ArgumentsResolverContextExtensions.HasArgument),
-            new[] { typeof(ResolverContextBase), typeof(string) }
-        )!;
-
     private static readonly MethodInfo ResolveValueTaskMethod = typeof(DelegateResolverFactory)
         .GetMethod(nameof(ResolveValueTask), BindingFlags.Static | BindingFlags.NonPublic)!;
 
@@ -53,48 +13,23 @@ public static class DelegateResolverFactory
 
     private static readonly MethodInfo ResolveValueObjectMethod = typeof(DelegateResolverFactory)
         .GetMethod(nameof(ResolveValueObject), BindingFlags.Static | BindingFlags.NonPublic)!;
+    
+    private static readonly Lazy<DelegateResolverFactory> InstanceFactory = new(() => new DelegateResolverFactory());
 
-    public static Resolver Create(Delegate resolverDelegate)
+    public static DelegateResolverFactory Instance => InstanceFactory.Value;
+
+    public static Resolver Get(Delegate resolverDelegate)
     {
-#if DEBUG
-        Trace.WriteLine(
-            $"Available context parameters:\n {string.Join(',', ContextParamProperties.Select(p => string.Concat($"{p.Key}: {p.Value.Type}")))}");
-#endif
+        return Instance.GetOrCreate(resolverDelegate);
+    }
 
+    public override Resolver Create(Delegate resolverDelegate)
+    {
         MethodInfo invokeMethod = resolverDelegate.Method;
 
-        Expression instanceExpression = null;
-        if (!invokeMethod.IsStatic) instanceExpression = Expression.Constant(resolverDelegate.Target);
-
-        IEnumerable<Expression> argumentsExpressions = invokeMethod.GetParameters()
-            .Select(p =>
-            {
-                if (p.ParameterType == typeof(ResolverContext)) return ContextParam;
-
-                if (p.Name is not null)
-                    if (ContextParamProperties.TryGetValue(p.Name.ToLowerInvariant(),
-                            out Expression? propertyExpression))
-                    {
-                        if (p.ParameterType == propertyExpression.Type)
-                            return propertyExpression;
-
-                        return Expression.Convert(propertyExpression, p.ParameterType);
-                    }
-
-                Expression hasArgumentCall = HasArgumentCall(p);
-                Expression resolveArgumentCall = ResolveArgumentCall(p);
-                Expression getRequiredServiceCall = GetRequiredServiceCall(p);
-                return Expression.Condition(
-                    hasArgumentCall,
-                    resolveArgumentCall,
-                    getRequiredServiceCall
-                );
-            });
-
-        MethodCallExpression invokeExpression = Expression.Call(
-            instanceExpression,
+        MethodCallExpression invokeExpression = GetDelegateMethodCallExpression(
             invokeMethod,
-            argumentsExpressions
+            resolverDelegate.Target
         );
 
         Expression valueTaskExpression;
@@ -144,104 +79,7 @@ public static class DelegateResolverFactory
         );
 
         Resolver compiledLambda = lambda.Compile();
-        Cache.TryAdd(resolverDelegate, compiledLambda);
         return compiledLambda;
-    }
-
-    public static Resolver GetOrCreate(Delegate resolverDelegate)
-    {
-        if (Cache.TryGetValue(resolverDelegate, out Resolver? resolver))
-            return resolver;
-
-        return Create(resolverDelegate);
-    }
-
-    private static Expression GetRequiredServiceCall(ParameterInfo p)
-    {
-        MemberExpression serviceProviderProperty = Expression.Property(ContextParam, "RequestServices");
-        MethodInfo? getServiceMethodInfo = typeof(ServiceProviderServiceExtensions)
-            .GetMethods()
-            .FirstOrDefault(m => m.Name == nameof(ServiceProviderServiceExtensions.GetRequiredService)
-                                 && m.GetParameters().Length == 1
-                                 && m.GetParameters()[0].ParameterType == typeof(IServiceProvider)
-                                 && m.IsGenericMethodDefinition);
-
-        if (getServiceMethodInfo is null)
-            throw new InvalidOperationException("Could not find GetRequiredService method");
-
-        Type serviceType = p.ParameterType;
-        MethodInfo genericMethodInfo = getServiceMethodInfo.MakeGenericMethod(serviceType);
-        var getRequiredServiceCall = (Expression)Expression.Call(
-            null,
-            genericMethodInfo,
-            serviceProviderProperty);
-        return getRequiredServiceCall;
-    }
-
-
-    private static Expression HasArgumentCall(ParameterInfo p)
-    {
-        ArgumentException.ThrowIfNullOrEmpty(p.Name);
-        ConstantExpression name = Expression.Constant(p.Name, typeof(string));
-        return Expression.Call(
-            null,
-            HasArgumentMethod,
-            ContextParam,
-            name);
-    }
-
-    private static bool IsPrimitiveOrNullablePrimitive(Type type)
-    {
-        if (type.IsPrimitive)
-            return true;
-        if (type.IsGenericType && type.GetGenericTypeDefinition() == typeof(Nullable<>))
-            return type.GetGenericArguments()[0].IsPrimitive;
-        return false;
-    }
-
-    private static Expression ResolveArgumentCall(ParameterInfo p)
-    {
-        ConstantExpression nameParam = Expression.Constant(p.Name, typeof(string));
-
-        Expression? callExpr;
-        if (IsPrimitiveOrNullablePrimitive(p.ParameterType))
-        {
-            MethodInfo methodInfo = GetArgumentMethod.MakeGenericMethod(p.ParameterType);
-            callExpr = Expression.Call(methodInfo, ContextParam, nameParam);
-        }
-        else if (p.ParameterType == typeof(string))
-        {
-            MethodInfo methodInfo = GetArgumentMethod.MakeGenericMethod(p.ParameterType);
-            callExpr = Expression.Call(methodInfo, ContextParam, nameParam);
-        }
-        else if (p.ParameterType.IsGenericType
-                 && p.ParameterType.GetGenericTypeDefinition() == typeof(IEnumerable<>)
-                 && (p.ParameterType.GetGenericArguments()[0].IsClass ||
-                     (p.ParameterType.GetGenericArguments()[0].IsValueType &&
-                      p.ParameterType.GetGenericArguments()[0].GetConstructor(Type.EmptyTypes) != null)))
-        {
-            MethodInfo methodInfo =
-                BindInputObjectListMethod.MakeGenericMethod(p.ParameterType.GetGenericArguments()[0]);
-            callExpr = Expression.Call(methodInfo, ContextParam, nameParam);
-        }
-        else if (p.ParameterType.IsClass)
-        {
-            MethodInfo methodInfo = BindInputObjectMethod.MakeGenericMethod(p.ParameterType);
-            callExpr = Expression.Call(methodInfo, ContextParam, nameParam);
-        }
-        else
-        {
-            MethodInfo methodInfo = ThrowMethod.MakeGenericMethod(p.ParameterType);
-            callExpr = Expression.Call(methodInfo,
-                Expression.Constant($"Unsupported parameter type '{p.ParameterType}' " +
-                                    $"for parameter '{p.Name}'. Only primitive types, classes " +
-                                    "and structs with a default constructor are supported."));
-        }
-
-
-        //callExpr = Expression.Convert(callExpr, p.ParameterType);
-
-        return callExpr;
     }
 
     private static ValueTask ResolveValueObject<T>(T result, ResolverContext context)
@@ -280,10 +118,5 @@ public static class DelegateResolverFactory
         }
 
         return AwaitResolveValue(task, context);
-    }
-
-    private static T? Throw<T>(string message)
-    {
-        throw new ArgumentException(message);
     }
 }
